Include column index tip in migration guide
I tried to use the migration guide to migrate a legacy app from PHPExcel, but I missed that column indices needed to be updated by hand. It's possible Rector is supposed to catch this and simply didn't in my case. 

This would reinstate that part of the migration guide so that future users know to update their column indices.